refactor(web): extract timeAgo helper for reuse (#129)
