+++ frontend/src/components/NotificationBell.jsx
@@ -1,13 +1,6 @@
 import { useState, useRef, useEffect, useCallback } from 'react'
 import { getNotifications, markNotificationAsRead, markAllNotificationsAsRead } from '../services/api'
-
-function timeAgo(iso) {
-  const diff = Math.floor((Date.now() - new Date(iso)) / 1000)
-  if (diff < 60) return 'just now'
-  if (diff < 3600) return `${Math.floor(diff / 60)}m ago`
-  if (diff < 86400) return `${Math.floor(diff / 3600)}h ago`
-  return `${Math.floor(diff / 86400)}d ago`
-}
+import { timeAgo } from '../utils/timeAgo'
 
 function TypeIcon({ type }) {
   if (type === 'REQUEST_ACCEPTED') {
